+++ Modules/Intent.Modules.Application.Dtos.Pagination/Templates/TemplateExtensions.cs
@@ -1,10 +1,10 @@
+using System.Collections.Generic;
 using Intent.Modules.Application.Dtos.Pagination.Templates.CursorPagedResult;
 using Intent.Modules.Application.Dtos.Pagination.Templates.CursorPagedResultMappingExtensions;
 using Intent.Modules.Application.Dtos.Pagination.Templates.PagedResult;
 using Intent.Modules.Application.Dtos.Pagination.Templates.PagedResultMappingExtensions;
 using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
-using System.Collections.Generic;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: DefaultIntentManaged(Mode.Fully, Targets = Targets.Usings)]
@@ -14,12 +14,12 @@ namespace Intent.Modules.Application.Dtos.Pagination.Templates
 {
     public static class TemplateExtensions
     {
-        public static string GetPagedResultName<T>(this IntentTemplateBase<T> template)
+        public static string GetPagedResultName(this IIntentTemplate template)
         {
             return template.GetTypeName(PagedResultTemplate.TemplateId);
         }
 
-        public static string GetCursorPagedResultName<T>(this IntentTemplateBase<T> template)
+        public static string GetCursorPagedResultName(this IIntentTemplate template)
         {
             return template.GetTypeName(CursorPagedResultTemplate.TemplateId);
         }
@@ -29,7 +29,7 @@ namespace Intent.Modules.Application.Dtos.Pagination.Templates
             return template.GetTypeName(CursorPagedResultMappingExtensionsTemplate.TemplateId);
         }
 
-        public static string GetPagedResultMappingExtensionsName<T>(this IntentTemplateBase<T> template)
+        public static string GetPagedResultMappingExtensionsName(this IIntentTemplate template)
         {
             return template.GetTypeName(PagedResultMappingExtensionsTemplate.TemplateId);
         }