+++ Modules/Intent.Modules.VisualStudio.Projects/OutputTargets/OutputTargetRegistration.cs
@@ -43,7 +43,7 @@ namespace Intent.Modules.VisualStudio.Projects.OutputTargets
                 var rootFolders = _metadataManager.CodebaseStructure(application).GetRootFolderModels();
                 foreach (var rootFolder in rootFolders)
                 {
-                    registry.RegisterOutputTarget(rootFolder.ToOutputTargetConfig());
+                    registry.RegisterOutputTarget(rootFolder.ToOutputTargetConfig(outputLocationOptions));
                     foreach (var f in rootFolder.Folders)
                     {
                         Register(registry, f);